Move NTP server dialog to its own file
